requirements.txt
absl-py==0.15.0
adal==1.2.7
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.8
albumentations==1.3.0
alembic==1.8.1
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.3.3
astunparse==1.6.3
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
azure-common==1.1.28
azure-core==1.25.0
azure-graphrbac==0.61.1
azure-mgmt-authorization==2.0.0
azure-mgmt-containerregistry==10.0.0
azure-mgmt-core==1.3.2
azure-mgmt-keyvault==10.1.0
azure-mgmt-resource==8.0.1
azure-mgmt-storage==20.1.0
azureml-core==1.0.85.6
Babel==2.6.0
backcall==0.1.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.11.1
black==22.3.0
bleach==3.1.1
blinker==1.4
blis==0.7.8
Boruta==0.3
cached-property==1.5.1
cachetools==5.2.0
catalogue==1.0.0
catboost==1.0.6
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==1.3.0
colorama==0.4.3
colorlover==0.3.0
configobj==5.0.6
constantly==15.1.0
contextlib2==21.6.0
cryptography==2.8
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.6
Cython==0.29.32
daal==2021.6.0
daal4py==2021.6.2
databricks-cli==0.17.1
dbus-python==1.2.16
debugpy==1.6.3
decorator==4.4.2
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
fairscale==0.4.6
fastapi==0.80.0
fastcache==1.1.0
filelock==3.8.0
Flask==2.1.2
Flask-RESTful==0.3.9
flatbuffers==1.12
fonttools==4.34.4
frozenlist==1.3.3
funcy==1.17
future==0.18.2
fvcore==0.1.5.post20220512
gast==0.4.0
gdown==4.5.1
gensim==3.8.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
greenlet==1.1.2
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
html5lib==1.0.1
htmlmin==0.1.12
httplib2==0.14.0
hydra-core==1.2.0
hyperlink==19.0.0
idna==2.8
ImageHash==4.2.1
imageio==2.21.1
imagesize==1.2.0
imbalanced-learn==0.7.0
imgaug==0.2.6
importlib-metadata==4.11.4
importlib-resources==5.9.0
incremental==16.10.1
iopath==0.1.9
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isodate==0.6.1
isort==4.3.4
itsdangerous==2.1.2
jedi==0.15.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonpatch==1.22
jsonpickle==2.2.0
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
jupyterlab-widgets==1.1.1
keras==2.8.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.4
kmodes==0.12.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
libclang==14.0.1
lightgbm==3.3.2
llvmlite==0.37.0
logilab-common==1.4.3
macaroonbakery==1.3.1
Mako==1.2.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
mccabe==0.6.1
missingno==0.5.1
mistune==0.8.4
mlflow==1.28.0
mlxtend==0.20.0
mock==3.0.5
more-itertools==4.2.0
msrest==0.7.1
msrestazure==0.6.4
multidict==6.0.2
multimethod==1.8
murmurhash==1.0.8
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.4
ndg-httpsclient==0.5.1
nest-asyncio==1.5.5
netifaces==0.10.4
nltk==3.7
notebook==6.4.12
numba==0.54.1
numexpr==2.8.3
numpy==1.19.5
numpydoc==0.7.0
oauthlib==3.1.0
olefile==0.46
omegaconf==2.2.2
opencv-python-headless==4.5.2.52
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandas-profiling==3.2.0
pandocfilters==1.4.2
parso==0.5.2
pathspec==0.9.0
patsy==0.5.2
pbr==5.4.5
pexpect==4.6.0
phik==0.12.2
pickleshare==0.7.5
Pillow==9.1.1
plac==1.1.3
platformdirs==2.5.2
plotly==5.9.0
portalocker==2.5.1
preshed==3.0.7
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.3
prompt-toolkit==2.0.10
protobuf==3.19.4
psutil==5.5.1
ptyprocess==0.7.0
py4j==0.10.9.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycocotools==2.0.4
pycodestyle==2.5.0
pycparser==2.21
pycups==1.9.73
pydantic==1.9.2
pydot==1.4.2
pyflakes==2.1.1
Pygments==2.3.1
pyOpenSSL==19.0.0
pyparsing==3.0.9
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
python-dateutil==2.8.1
PyWavelets==1.3.0
PyYAML==5.1
QtPy==1.9.0
regex==2022.7.25
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
roman==2.0.0
rope==0.10.5
rsa==4.8
ruamel.yaml==0.15.89
scikit-image==0.18.3
scikit-learn==1.0.2
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.8.0
service-identity==18.1.0
shap==0.41.0
six==1.15.0
sklearn==0.0
sklearn-pandas==1.8.0
slicer==0.0.7
smart-open==6.0.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==2.3.7
Sphinx==1.8.5
spyder==3.3.6
spyder-kernels==0.5.2
ssh-import-id==5.10
starlette==0.19.1
statsmodels==0.13.2
tabulate==0.8.10
tenacity==8.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
testpath==0.4.4
textblob==0.17.1
texttable==1.6.2
thinc==7.4.5
threadpoolctl==3.1.0
timm==0.6.7
tomli==2.0.1
tornado==6.2
tqdm==4.64.0
traitlets==4.3.3
typing-extensions==4.4.0
urllib3==1.25.8
visions==0.7.4
wadllib==1.3.3
wasabi==0.10.1
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==2.1.2
widgetsnbextension==3.6.1
wordcloud==1.8.2.2
wrapt==1.11.*
xgboost==0.90
yacs==0.1.8
yarl==1.8.1
yellowbrick==1.4
zipp==3.8.1
zope.interface==4.7.1